Making the two display interface independent
issues when enabling the picture display interface and issue when resizing the picture display interface

+++ src/gui/AutoscopeWindowForm.hpp
@@ -88,7 +88,7 @@ private slots:
     void displaySizeChanged(int);
     void horizontalDisplayPositionChanged(int);
     void verticalDisplayPositionChanged(int);
-    void displayOpacityChanged(double);
+    void displayOpacityChanged(int);
     void outputPictureDirectoryChanged(void);
     void outputPictureDirectoryButtonPressed(void);
     void downloadPictureButtonPressed(void);
